+++ src/RFC4180Field.php
@@ -168,6 +168,7 @@ class RFC4180Field extends php_user_filter
     }
 
     /**
+     * @codeCoverageIgnore
      * Validate params property.
      */
     protected function isValidParams(array $params): bool